Move deku_helpers and connection_token into their own file
